requirements.txt
fastapi==0.115.11         # API framework
uvicorn==0.33.0
pyjwt==2.9.0
redis==5.2.1
python-dotenv==1.0.1     # Load environment variables
httpx==0.28.1
langchain==0.2.17
langchain-community==0.2.19
langchainhub==0.1.21
groq==0.18.0
google-genai==1.7.0
